Refactor to use listeners
