+++ esphome/components/sdl/sdl_esphome.cpp
@@ -61,6 +61,12 @@ void Sdl::draw_pixel_at(int x, int y, Color color) {
     this->y_high_ = y;
 }
 
+void Sdl::process_key(uint32_t keycode, bool down) {
+  auto callback = this->key_callbacks_.find(keycode);
+  if (callback != this->key_callbacks_.end())
+    callback->second(down);
+}
+
 void Sdl::loop() {
   SDL_Event e;
   if (SDL_PollEvent(&e)) {
@@ -88,11 +94,13 @@ void Sdl::loop() {
         break;
 
       case SDL_KEYDOWN:
-        this->key_state[e.key.keysym.sym] = true;
+        ESP_LOGD(TAG, "keydown %d", e.key.keysym.sym);
+        this->process_key(e.key.keysym.sym, true);
         break;
 
       case SDL_KEYUP:
-        this->key_state[e.key.keysym.sym] = false;
+        ESP_LOGD(TAG, "keyup %d", e.key.keysym.sym);
+        this->process_key(e.key.keysym.sym, false);
         break;
 
       case SDL_WINDOWEVENT:
